Fix the bug that stream is closed with wrong error code
This commit fixes the bug that stream is closed with wrong error code
(0).  This happens when STREAM or DATA frame with END_STREAM flag set
is received and it violates HTTP messaging rule (i.e., content-length
does not match).  In this case, nghttp2_on_stream_close_callback
should be called with nonzero error code, but previously it is called
with 0 (NO_ERROR).